Flytter oppgaveRequestMedDefaultEnhet til utils for OppgaveRequest
