+++ src/frontend/Sider/Oppgavebenk/filter/filterutils.ts
@@ -1,4 +1,3 @@
-import { FortroligEnhet, IkkeFortroligEnhet } from '../typer/enhet';
 import { OppgaveRequest } from '../typer/oppgave';
 
 export const oppdaterFilter = (
@@ -16,22 +15,4 @@ export const oppdaterFilter = (
         ...object,
         [key]: val,
     };
-};
-
-export const oppgaveRequestMedDefaultEnhet = (
-    oppgaveRequest: OppgaveRequest,
-    harSaksbehandlerStrengtFortroligRolle: boolean
-): OppgaveRequest => {
-    if (harSaksbehandlerStrengtFortroligRolle) {
-        return {
-            ...oppgaveRequest,
-            enhet: FortroligEnhet.VIKAFOSSEN,
-        };
-    } else {
-        const enhet = oppgaveRequest.enhet;
-        return {
-            ...oppgaveRequest,
-            enhet: enhet || IkkeFortroligEnhet.NAY,
-        };
-    }
 };
